Breaking out Region and Image.
