+++ dev/UltimaXNA.csproj
@@ -197,6 +197,10 @@
     <Compile Include="Core\Resources\IAnimationFrame.cs" />
     <Compile Include="Core\Resources\IResource.cs" />
     <Compile Include="Core\UI\ControlMetaData.cs" />
+    <Compile Include="Core\UI\HTML\Images.cs" />
+    <Compile Include="Core\UI\HTML\Layout\Box.cs" />
+    <Compile Include="Core\UI\HTML\Layout\LayoutTree.cs" />
+    <Compile Include="Core\UI\HTML\Region.cs" />
     <Compile Include="Core\UI\HTML\Styles\OpenTag.cs" />
     <Compile Include="Core\UI\HTML\Styles\StyleManager.cs" />
     <Compile Include="Core\UI\HTML\HREFAttributes.cs" />
@@ -501,9 +505,9 @@
     <Compile Include="Ultima\Resources\SpeechData.cs" />
     <Compile Include="Ultima\Resources\TileMatrixData.cs" />
     <Compile Include="Ultima\World\Input\WorldInput.cs" />
-    <Compile Include="Core\UI\HTML\Regions.cs" />
+    <Compile Include="Core\UI\HTML\RegionList.cs" />
     <Compile Include="Core\UI\HTML\Styles\Enums.cs" />
-    <Compile Include="Core\UI\HTML\Images.cs" />
+    <Compile Include="Core\UI\HTML\ImageList.cs" />
     <Compile Include="Core\UI\HTML\Atoms\Atomizer.cs" />
     <Compile Include="Core\Diagnostics\Metrics.cs" />
     <Compile Include="Ultima\Resources\GumpMulResource.cs" />
@@ -682,9 +686,7 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Core\UI\HTML\Layout\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <Import Project="$(MSBuildExtensionsPath)\Microsoft\XNA Game Studio\Microsoft.Xna.GameStudio.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
